+++ flex_gemm/kernels/triton/spconv/sparse_submanifold_conv_fwd_implicit_gemm_splitk.py
@@ -48,7 +48,7 @@ def sparse_submanifold_conv_fwd_implicit_gemm_splitk_kernel(
     num_k = tl.cdiv(Ci, BK)  # Number of blocks in K dimension
     k_start = tl.cdiv(num_k * V * block_id_k, SPLITK)
     k_end = tl.cdiv(num_k * V * (block_id_k + 1), SPLITK)
-    offset_n = (block_id_n * B1 + tl.arange(0, B1)) % N         # (B1,)
+    offset_n = (block_id_n.to(tl.int64) * B1 + tl.arange(0, B1)) % N         # (B1,)
     offset_co = (block_id_co * B2 + tl.arange(0, B2)) % Co      # (B2,)
     offset_k = tl.arange(0, BK)                                 # (BK,)
     
@@ -63,7 +63,7 @@ def sparse_submanifold_conv_fwd_implicit_gemm_splitk_kernel(
         v = k // num_k
         bk = k % num_k
         # Calculate pointers to input matrix.
-        neighbor_offset_n = tl.load(neighbor + offset_n * V + v).to(tl.int64)                   # (B1,)
+        neighbor_offset_n = tl.load(neighbor + offset_n.to(tl.int64) * V + v).to(tl.int64)                   # (B1,)
         input_ptr = input + bk * BK + (neighbor_offset_n[:, None].to(tl.int64) * Ci + offset_k[None, :])     # (B1, BK)
         # Load the next block of input and weight.
         neigh_mask = neighbor_offset_n != 0xffffffff
@@ -82,9 +82,9 @@ def sparse_submanifold_conv_fwd_implicit_gemm_splitk_kernel(
         accumulator += bias_block[None, :]
                 
     # Write back the block of the output matrix with masks.
-    out_offset_n = block_id_n * B1 + tl.arange(0, B1)
+    out_offset_n = block_id_n.to(tl.int64) * B1 + tl.arange(0, B1)
     out_offset_co = block_id_co * B2 + tl.arange(0, B2)
-    out_ptr = output + block_id_k * N * Co + (out_offset_n[:, None] * Co + out_offset_co[None, :])
+    out_ptr = output + block_id_k.to(tl.int64) * N * Co + (out_offset_n[:, None] * Co + out_offset_co[None, :])
     out_mask = (out_offset_n[:, None] < N) & (out_offset_co[None, :] < Co)
     tl.store(out_ptr, accumulator, mask=out_mask)
